test: split test-cpu-prof.js
Split test-cpu-prof.js into multiple files for different
test cases so it's easier to find the problematic one if
it flakes.

Also move the split tests into parallel.

PR-URL: https://github.com/nodejs/node/pull/28170
Refs: https://github.com/nodejs/node/issues/27611
